pyproject.toml
[tool.poetry]
name = "bitrix-helper"
version = "0.1.0"
description = ""
authors = ["Igor Gerasimov <gerasimov.igor.98@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
selenium = "^4.25.0"
webdriver-manager = "^4.0.2"
requests = "^2.32.3"
aiogram = "^3.13.1"
tqdm = "^4.66.5"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
telebot = "^0.0.5"
chromadb = "^0.5.11"
langchain-community = "^0.3.1"
openai = "^1.51.0"
loguru = "^0.7.2"
redis = "^5.1.1"
tiktoken = "^0.8.0"
sentence-transformers = "^3.1.1"
beautifulsoup4 = "^4.12.3"
fast-bitrix24 = "^1.8.0"
pytz = "^2024.2"
langchain-ollama = "^0.2.0"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
google-auth = "^2.36.0"
google-api-python-client = "^2.151.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
certifi = "^2024.8.30"
pybitrix24 = "^1.1.0"
boto3 = "^1.35.60"
pytube = "^15.0.0"
pydub = "^0.25.1"
yt-dlp = "^2024.11.4"
moviepy = "^1.0.3"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
